% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DBIReporter.R
\docType{class}
\name{DBIReporter-class}
\alias{DBIReporter}
\alias{DBIReporter-class}
\title{Class record testthat tests into a database.}
\arguments{
\item{dbh}{the database handle}

\item{...}{Arguments used to initialise class}
}
\description{
This is an extension to testthat Reporter class.
}
\examples{
dbh <- dbConnect(RSQLite::SQLite(), ':memory:')
DBIReporter$new(dbh, 'test_results')

This reporter will gather all test results and insert them into an SQLite
database given as a open handled upoon initialization.
}
\keyword{debugging}

